requirements.txt
matplotlib
numpy
pandas
polars
pyyaml
torch
scipy
scikit-learns